+++ Flow/src/utopia/flow/collection/template/MatrixViewLike.scala
@@ -8,16 +8,24 @@ import utopia.flow.collection.immutable.range.NumericSpan
   * @author Mikko Hilpinen
   * @since 23.1.2023, v2.0
   * @tparam A Type of matrix cell values
+  * @tparam O Type of values in the viewed matrix
   * @tparam M Type of this view as a matrix
   */
-trait MatrixViewLike[+A, +M] extends MatrixLike[A, M]
+trait MatrixViewLike[+A, O, +M] extends MatrixLike[A, M]
 {
 	// ABSTRACT ----------------------------
 	
 	/**
 	  * @return The matrix viewed through this view
 	  */
-	protected def matrix: MatrixLike[A, _]
+	protected def matrix: MatrixLike[O, _]
+	/**
+	  * @return A mapping function applied to the viewed matrix values.
+	  *         Either Left: A simple mapping function, or
+	  *         Right: An index-based mapping function that receives both the original value
+	  *         and its index within this view
+	  */
+	protected def viewFunction: Either[O => A, (O, Pair[Int]) => A]
 	
 	/**
 	  * @return The viewed sub-region of the viewed matrix.
@@ -62,20 +70,43 @@ trait MatrixViewLike[+A, +M] extends MatrixLike[A, M]
 	  */
 	def endY = verticalViewRange.end
 	
+	/**
+	  * @return View to the columns in the original matrix
+	  */
+	protected def originalColumnsView =
+		matrix.columnsView.slice(horizontalViewRange.start, horizontalViewRange.end + 1)
+			.map { _.slice(verticalViewRange.start, verticalViewRange.end + 1) }
+	/**
+	  * @return View to the rows in the original matrix
+	  */
+	protected def originalRowsView =
+		matrix.rowsView.slice(verticalViewRange.start, verticalViewRange.end + 1)
+			.map { _.slice(horizontalViewRange.start, horizontalViewRange.end + 1) }
+	
 	
 	// IMPLEMENTED  ------------------------
 	
 	override def width = horizontalViewRange.length + 1
 	override def height = verticalViewRange.length + 1
 	
-	override def columnsView =
-		matrix.columnsView.slice(horizontalViewRange.start, horizontalViewRange.end + 1)
-			.map { _.slice(verticalViewRange.start, verticalViewRange.end + 1) }
-	override def rowsView =
-		matrix.rowsView.slice(verticalViewRange.start, verticalViewRange.end + 1)
-			.map { _.slice(horizontalViewRange.start, horizontalViewRange.end + 1) }
+	override def columnsView = viewFunction match {
+		case Left(f) => originalColumnsView.map { _.map(f) }
+		case Right(f) =>
+			(0 until width).view.map { x => (0 until height).view.map { y => f(original(x, y), Pair(x, y)) } }
+	}
+	override def rowsView = viewFunction match {
+		case Left(f) => originalRowsView.map { _.map(f) }
+		case Right(f) =>
+			(0 until height).view.map { y => (0 until width).view.map { x => f(original(x, y), Pair(x, y)) } }
+	}
 	
-	override def apply(column: Int, row: Int) = matrix(startX + column, startY + row)
+	override def apply(column: Int, row: Int) = {
+		val value = original(column, row)
+		viewFunction match {
+			case Left(f) => f(value)
+			case Right(f) => f(value, Pair(column, row))
+		}
+	}
 	
 	override def view(area: Pair[NumericSpan[Int]]): M = viewArea.findMergeWith(area) { _ overlapWith _ } match {
 		case Some(overlap) => withViewArea(overlap)
@@ -97,6 +128,18 @@ trait MatrixViewLike[+A, +M] extends MatrixLike[A, M]
 	
 	
 	// OTHER    --------------------------
+	
+	/**
+	  * @param column Targeted column index (x)
+	  * @param row Targeted row index (y)
+	  * @return Cell value of the viewed matrix at that location
+	  */
+	protected def original(column: Int, row: Int) = matrix(startX + column, startY + row)
+	/**
+	  * @param point Targeted point within this matrix view
+	  * @return Cell value of the viewed matrix at that location
+	  */
+	protected def original(point: Pair[Int]) = matrix(viewArea.mergeWith(point) { _.start + _ })
 	
 	/**
 	  * @param amount Amount to shift this view horizontally (first) and vertically (second)